requirements.txt
python-binance
websocket-client
python-dotenv~=1.1.1
tabulate~=0.9.0


pandas~=2.3.1
numpy~=2.0.2
tqdm~=4.67.1